requirements.txt
cryptography==36.0.2
Django==3.1.3
django-filter==21.1
django-ipware==4.0.2
django-utils-six==2.0
djangorestframework==3.13.1
oauthlib==3.2.0
Pillow==8.0.1
PyJWT==2.3.0
requests==2.27.1
requests-oauthlib==1.3.1